moving render complete to embeddable
